+++ NEWS.md
@@ -1,8 +1,9 @@
 
-# RcppDynProg 0.1.1 2019/01/29
+# RcppDynProg 0.1.1 2019/01/31
 
  * Add logistic scoring.
  * No k-bound variation.
+ * Add a logistic fitter.
 
 # RcppDynProg 0.1.0 2019/01/01